Some refactor for icons
