+++ src/components/mapgl-legend/mapgl-legend.component.ts
@@ -6,7 +6,6 @@ import { scaleLinear, ScaleLinear } from 'd3-scale';
 import { select } from 'd3-selection';
 import { HistogramData } from 'arlas-d3/histograms/utils/HistogramUtils';
 import { StyleFunction, Expression } from 'mapbox-gl';
-import { ColorLegend, PROPERTY_SELECTOR_SOURCE, Legend } from './legend';
 
 
 export const GET = 'get';
@@ -18,24 +17,23 @@ export const OTHER = 'other_color';
   templateUrl: './mapgl-legend.component.html',
   styleUrls: ['./mapgl-legend.component.css']
 })
-export class MapglLegendComponent extends ColorLegend implements OnInit, AfterViewInit, OnChanges {
+export class MapglLegendComponent implements OnInit, AfterViewInit, OnChanges {
 
   @Input() public layer: mapboxgl.Layer;
   @Input() public legendUpdater: Subject<Map<string, {minValue: string, maxValue: string}>> =
     new Subject<Map<string, {minValue: string, maxValue: string}>>();
   @ViewChild('width_svg', { read: ElementRef, static: false }) public lineWidthLegendElement: ElementRef;
   @ViewChild('radius_svg', { read: ElementRef, static: false }) public circleRadiusLegendElement: ElementRef;
-
+  public colorLegend: Legend = {};
   public widthLegend: Legend = {};
   public radiusLegend: Legend = {};
   public detail = false;
+  public PROPERTY_SELECTOR_SOURCE = PROPERTY_SELECTOR_SOURCE;
   private legendData: Map<string, {minValue: string, maxValue: string}> = new Map();
 
   private MAX_LINE_WIDTH = 10;
   private MAX_CIRLE_RADIUS = 7;
-  constructor(public translate: TranslateService, private el: ElementRef) {
-    super(translate);
-   }
+  constructor(public translate: TranslateService, private el: ElementRef) {}
 
   public ngOnInit() {
     this.legendUpdater.subscribe(legendData => {
@@ -94,6 +92,44 @@ export class MapglLegendComponent extends ColorLegend implements OnInit, AfterVi
   }
 
 
+  private buildColorLegend(color: string | StyleFunction | Expression, legendData?: any): void {
+    if (typeof color === 'string') {
+      this.colorLegend.type = PROPERTY_SELECTOR_SOURCE.fix;
+    } else if (Array.isArray(color)) {
+      if (color.length === 2) {
+        /** color = ["get", "field"]  ==> Generated or Provided */
+        // todo
+      } else if (color.length >= 3) {
+        if (color[0] === MATCH) {
+          /** color = ["match", ["get", "field"], .... ]**/
+          this.colorLegend.type = PROPERTY_SELECTOR_SOURCE.manual;
+          const colorsLength = color.length - 2;
+          let hasDefaultColor = false;
+          if (colorsLength % 2 !== 0) {
+            hasDefaultColor = true;
+          }
+          for (let i = 2; i < color.length; i += 2) {
+            if (hasDefaultColor && i === colorsLength - 1) {
+              this.colorLegend.manualValues.set(this.translate.instant(OTHER), color[i]);
+            } else {
+              this.colorLegend.manualValues.set(this.translate.instant(color[i]), color[i + 1]);
+            }
+          }
+        } else if (color[0] === INTERPOLATE) {
+          this.colorLegend.type = PROPERTY_SELECTOR_SOURCE.interpolated;
+          /** color = ["interplate", ['linear'], ["get", "field"], 0, 1... ]**/
+          // todo throw exception if interpolation is not linear
+          const field = color[2][1];
+          this.colorLegend.title = field;
+          if (legendData && legendData.get(field)) {
+            this.colorLegend.minValue = legendData.get(field).minValue;
+            this.colorLegend.maxValue = legendData.get(field).maxValue;
+          }
+          this.colorLegend.interpolatedValues = color.filter((c, i) => i > 2 && i % 2 === 0);
+        }
+      }
+    }
+  }
 
   private buildLineWidthLegend(lineWidth: number | StyleFunction | Expression): void {
     if (Array.isArray(lineWidth)) {
@@ -215,4 +251,24 @@ export function drawCircleSupportLine(svgNode: SVGElement, circlesRadiuses: Arra
 
 export function getMax(data: Array<HistogramData>): number {
   return Math.max(...data.map(hd => +hd.value));
+}
+
+export interface Legend {
+  type?: PROPERTY_SELECTOR_SOURCE;
+  title?: string;
+  minValue?: string;
+  maxValue?: string;
+  fixValue?: string | number;
+  interpolatedValues?: Array<string> | Array<number>;
+  manualValues?: Map<string, string | number>;
+}
+
+export enum PROPERTY_SELECTOR_SOURCE {
+  fix = 'Fix',
+  provided = 'Provided',
+  generated = 'Generated',
+  manual = 'Manual',
+  interpolated = 'Interpolated',
+  metric_on_field = 'Metric on field',
+  heatmap_density = 'Density'
 }
